Adding a bunch of Mac-side control bindings
In doing some of my own work, I found that I was duplicating a lot of what was already done in Rex, so I ported over much of what I've done on my own in order to better consolidate things.

There isn't a complete 1:1 test matching at the moment with the new AppKit methods, and I "broke" the API for NSTextField that was already in place to better match with conventions used elsewhere in RAC, and to distinguish between continuous and "normal" control events.