Adding my work on control bindings
Originally developed alongside a forthcoming update to FuzzMeasure
(http://fuzzmeasure.com), this is my take on one way that Cocoa
Bindings-like behavior may be approximated using RAC.

In addition to the `Signal`s that contain controls' changing values, the
`ConsumerBinding` provides the "other half" of the equation—updating
their state to reflect incoming values. Like Cocoa Bindings, values are
wrapped by placeholders that represent selection states such as
`MultipleValues` and `NoSelection`.

A new `PropertyBinding` is a similar concept that exists for model
objects to expose, where a `validator` function is provided to allow the
"massaging" of incoming values. This approximates the role of Key-Value
Validation as it existed within Cocoa Bindings.

This changeset does not include any examples, sadly. Hopefully I can put
something together in the near future to share.