Add bindable property to wrap a control's value
This property is common to every `UIControl` and can be used by
providing a getter and a setter for the value wrapped. This property
uses `UIControlEvents.ValueChanged` and `UIControlEvents.EditingChanged`
events to detect changes and keep the value up-to-date. This kind of
logic can be reused instead of being defined and used for each control.